Add files via upload
Here we perform multiple sequence alignment, pairwise sequence alignment and use a heatmap for visualization of the pairwise sequence alignments to analyze the similarity between the human, tree shrew, and mouse TRPV1 sequences. Multiple sequence alignment is conducted to print out and visualize the alignment of all three TRPV1 sequences. Pairwise sequence alignment is used to compare the similarity between pairs of TRPV1 sequences, and a heatmap is used to to visualize the results of the pairwise alignments.

Scientific Question: Is the TRPV1 sequence of humans more similar to the TRPV1 sequence of tree shrews or the TRPV1 sequence of mice?
Scientific Hypothesis: If humans actively seek spicy foods in a similar fashion to tree shrews, then the TRPV1 sequence of humans is more likely to have a higher similarity to the TRPV1 sequence of tree shrews than the TRPV1 sequence of mice.